mad_spark_multiagent/requirements.txt
google-adk
google-generativeai
python-dotenv

# Testing dependencies
pytest>=8.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
